bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Stanislav <sta-tun-slav@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiogram = "^3.12.0"
tortoise-orm = "^0.21.6"
asyncpg = "^0.29.0"
fastapi = {extras = ["standard"], version = "^0.112.2"}

[tool.poetry.scripts]
start = "src.bot:start"
init_db = "src.db.db:run_init_db"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
